# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MeetExpert API",
        "description": "MeetExpertBD REST API documentation (auth, user profile, experts, bookings).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/v1/expert/application": {
            "post": {
                "tags": [
                    "Expert Application"
                ],
                "summary": "Submit an expert application",
                "operationId": "a1e6555478e90f2c7146171e29909a23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "subcategory_id",
                                    "professional_headline",
                                    "bio",
                                    "years_of_experience",
                                    "registration_value",
                                    "languages",
                                    "skill_ids"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "subcategory_id": {
                                        "type": "integer"
                                    },
                                    "professional_headline": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "bio": {
                                        "type": "string",
                                        "maxLength": 10000
                                    },
                                    "years_of_experience": {
                                        "type": "integer",
                                        "maximum": 80,
                                        "minimum": 0
                                    },
                                    "registration_value": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "intro_video": {
                                        "type": "string",
                                        "format": "uri",
                                        "nullable": true,
                                        "maxLength": 2048
                                    },
                                    "languages": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "maxLength": 50
                                        },
                                        "maxItems": 20,
                                        "minItems": 1
                                    },
                                    "education": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "institution"
                                            ],
                                            "properties": {
                                                "institution": {
                                                    "type": "string",
                                                    "maxLength": 255
                                                },
                                                "degree": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 255
                                                },
                                                "year": {
                                                    "type": "integer",
                                                    "nullable": true,
                                                    "maximum": 2100,
                                                    "minimum": 1900
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true,
                                        "maxItems": 20
                                    },
                                    "experience": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "title"
                                            ],
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "maxLength": 255
                                                },
                                                "organization": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 255
                                                },
                                                "start_year": {
                                                    "type": "integer",
                                                    "nullable": true,
                                                    "maximum": 2100,
                                                    "minimum": 1900
                                                },
                                                "end_year": {
                                                    "type": "integer",
                                                    "nullable": true,
                                                    "maximum": 2100,
                                                    "minimum": 1900
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 2000
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true,
                                        "maxItems": 30
                                    },
                                    "portfolio": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "url"
                                            ],
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "nullable": true,
                                                    "maxLength": 255
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "maxLength": 2048
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true,
                                        "maxItems": 20
                                    },
                                    "skill_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "maxItems": 50,
                                        "minItems": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter Sanctum personal access token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}